Resync with upstream and reenable system-wide ffmpeg linking support.
Was disabled upstream for an unknown reason.

(Can't put the code in an else statement block, make the if statement it's attached to always true before the condition is checked, and expect the code to execute.) (It only does something if the user defines some undocumented cmake options. So really it's hidden anyway, unless the user looks for it.)

Also reenabled the ffmpeg library location messages. (Just so if debugging is needed, we know what ffmpeg we are using. (system or internal.))

Primary reason for this commit: https://github.com/hrydgard/ppsspp/issues/5772
(Also I do have a spec file for Mageia, it's why I made this patch in the first place.)