Add the ability to set block_align when constructing a SimpleAudio object.
This commit modifies SimpleAudio's constructor and Init() functions to allow setting AVCodecContext::block_align when a SimpleAudio object is created. (Default is to set it to zero to keep the original behavior if not needed.)

This seems to be required now with newer versions of ffmpeg (version 2.2+) to init the atrac3+ decoder, as avcodec_open2() will return -22 if it's not set / set to zero.