Hard-code a block_align value for SimpleAudio in MediaEngine::openContext()
This allows some games (Corpse Party, and Crisis Core were tested) to have working atrac3+ audio.
(The value is 280.)

This value should be computed / pulled from the audio stream, but I don't know enough about the MediaEngine to find where this info could be kept. Some input would be appreciated.

Does not truly fix issue #5772, but it's somewhat a step in the right direction. (If not a hacky one. :P)